+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 ### Removed
 
 ### Fixed
+- Potential deadlock in `Lazy` when accessed for the first time from
+  multiple threads simultaneously.
 
 ### Security